Add ability to make async blocks sync
There are a lot of cases this won't work for but at a minimum we can now
turn async blocks that look like this:

```
let x = async move { some_future().await };
x.await?;
```

into

```
let x = { some_future() };
x?;
```

This, however, is a really basic transformation and doesn't account for
things like becoming pinned, being spawned as is etc. it might be worth
changing this to a closure to better capture the semantics of what it's
doing, though it's hard to say what's better to implement without enough
usage.