Update Cargo.toml to have data for publishing
The Cargo.toml file didn't have everything necessary for it to be
published to crates.io. This commit rectifies that by adding it and we
set the version to 0.0.1 as getting people to be able to use this would
be great for figuring out where kinks are and then having issues filed
to start sanding down corner cases. We also want it to be this version
as it's very clearly not ready for production use cases yet at all.